pyproject.toml
[tool.poetry]
name = "semantic-ner-search-demo"
version = "0.1.0"
description = "Python Poetry Boilerplate"
authors = ["face0b1101 <6190957+face0b1101@users.noreply.github.com>"]
maintainers = ["face0b1101 <6190957+face0b1101@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://github.com/face0b1101/semantic-ner-search-demo"
repository = "https://github.com/face0b1101/semantic-ner-search-demo"
documentation = "https://github.com/face0b1101/semantic-ner-search-demo"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
unipath = "^1.1"
rich = "^13.7.0"
tqdm = "^4.66.1"
eland = "^8.12.1"
elasticsearch = "^8.12.1"
streamlit = "^1.32.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
jupyter = "^1.0.0"
pytest = "^7.4.4"
isort = "^5.13.2"
tox = "^4.11.4"
flake8-bugbear = "^23.12.2"
kaggle = "^1.6.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
# make it compatible with black
profile = "black"

[tool.poetry.scripts]
semantic-ner-search-demo = "semantic_ner_search_demo.app:main"
